# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin_desa
        type: string
    required:
    - password
    - username
    type: object
...
  auth.RequestUserFormat:
    properties:
      password:
        example: rahasia123
        type: string
      roleId:
        description: <-- Harus UUID dari tabel roles
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      username:
        example: kepala_desa_1
        type: string
    required:
    - roleId
    - username
    type: object
  master.RequestKriteriaFormat:
    properties:
      bobot:
        type: number
      id:
        type: string
      jenis:
        enum:
        - benefit
        - cost
        type: string
      kode:
        description: Tambahan wajib diisi
        example: C1
        type: string
      nama:
        type: string
    required:
    - bobot
    - jenis
    - kode
    - nama
    type: object
...
  response.Base:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Backend untuk Sistem Pendukung Keputusan (SIPROPEDA) - Metode TOPSIS.
  title: SIPROPEDA API
  version: "1.0"
paths:
  /v1/kriteria:
    get:
      description: Endpoint untuk mengambil daftar kriteria yang aktif
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil Semua Data Kriteria
      tags:
      - Kriteria
    post:
      consumes:
      - application/json
      description: Endpoint ini digunakan untuk menambahkan Kriteria baru ke database.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Kriteria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/master.RequestKriteriaFormat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah data Kriteria baru
      tags:
      - Kriteria
  /v1/kriteria/{id}:
    delete:
      description: Endpoint ini digunakan untuk menghapus data Kriteria berdasarkan
        ID (soft delete).
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Kriteria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus Kriteria by ID (Soft Delete)
      tags:
      - Kriteria
    get:
      description: Endpoint ini digunakan untuk mengambil satu data Kriteria berdasarkan
        ID.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Kriteria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail Kriteria by ID
      tags:
      - Kriteria
    put:
      consumes:
      - application/json
      description: Endpoint ini digunakan untuk memperbarui data Kriteria berdasarkan
        ID yang dikirimkan di path.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Kriteria
        in: path
        name: id
        required: true
        type: string
      - description: Data Kriteria yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/master.RequestKriteriaFormat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Base'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Base'
      summary: Perbarui data Kriteria
      tags:
      - Kriteria
  /v1/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Kredensial Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Base'
      summary: Login User
      tags:
      - Auth
...
  /v1/user:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil Semua Data User
      tags:
      - User
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestUserFormat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah User Baru
      tags:
      - User
  /v1/user/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus User by ID (Soft Delete)
      tags:
      - User
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail User by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID User
        in: path
        name: id
        required: true
        type: string
      - description: Data User yang diperbarui (kosongkan password jika tidak ingin
          ganti)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestUserFormat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Perbarui data User
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"